+++ src/test/java/racingcar/CarNameTest.java
@@ -1,7 +1,8 @@
-package racingcar.domain;
+package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.CarName;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
