+++ src/test/java/racingcar/CarRepositoryTest.java
@@ -1,7 +1,8 @@
-package racingcar.domain;
+package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.CarRepository;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
